requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
passlib[bycrypt]
python-jose[crytography]